requirements.txt
# Core API
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Image Processing & OCR
pillow>=10.0.0
pytesseract>=0.3.10
opencv-python-headless>=4.8.0  # headless for Docker (no GUI)
easyocr>=1.7.0

# HTTP & Utilities
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
openai>=1.0.0
httpx>=0.25.0

# AI Agent (LangGraph + Ollama)
langgraph>=0.2.0
langchain>=0.2.0
langchain-ollama>=0.1.0
langchain-community>=0.2.0

# Production Database
sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.0  # PostgreSQL (uncomment if needed)
